assistant/requirements.txt
python-vlc
flask